Modbus串口错误发送处理

+++ DeviceStatus/RealTimeCurve/rtcurve.cpp
@@ -964,7 +964,7 @@ void RTCurve::data_process(const QModbusDataUnit unit)
             }
             break;
         case InputRegs_TT_28:
-            value[0] = unit.value(i);
+            value[0] = qint16(unit.value(i));
 
             values[TT25_TT32].push_back(unit.value(i));
